pyproject.toml
[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "pyopenxlsx"
version = "0.2.0"
description = "Python bindings for OpenXLSX using pybind11"
authors = [{name = "curry tang", email = "twn39@163.com"}]
license = {text = "MIT"}
requires-python = ">=3.11"

[tool.scikit-build]
cmake.version = ">=3.15"
wheel.packages = ["src/pyopenxlsx"]

[dependency-groups]
dev = [
    "clang-tidy>=21.1.6",
    "openpyxl>=3.1.5",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
    "ty>=0.0.15",
    "pytest-benchmark>=4.0.0",
    "numpy>=1.26.0",
]

[tool.pytest.ini_options]
addopts = "--cov=pyopenxlsx --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
norecursedirs = ["third_party"]

[tool.coverage.run]
source = ["src/pyopenxlsx"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.ty.analysis]
allowed-unresolved-imports = ["pyopenxlsx._openxlsx"]
